Switch to font-lock-default-fontify-region
By this time, the only thing ‘ledger-fontify-buffer-part’ does is ensure
that ‘font-lock-fontify-keywords-region’ is called for regions that do
not begin or end in the middle of a transaction or directive.

This can be achieved by setting an appropriate
‘font-lock-extend-region-function’.

* ledger-fontify-buffer-part: Remove.
* ledger-fontify-extend-region: Extend region to transaction and
  directive boundaries.
* ledger-mode: Do not set ‘font-lock-fontify-region-function’; set
  ‘font-lock-extend-region-functions’.
* ledger-navigate-find-directive-extents: Do not return positions beyond
  the buffer, ‘font-lock-default-fontify-region’ does not like that.

After this change, ledger-fontify/test-089 fails because the transaction
start regexp got relaxed to “starts with a digit” which should be better
for interactive use. This is not a big deal ’cause we’re going to handle
block comments soon.