Fontify arguments of account directive and subdirectives
Previously, the whole line got fontified. Now, we try to highlight parts
of directives according to their semantics.

A naïve implementation would use a regex that matches valid text and
rejects invalid or incomplete text. However, in syntax highlighting, it
is desirable to highlight incomplete text as it would be highlighted
when completed.

Therefore, we add tests for incomplete text. There are two things we’re
checking for:

* incomplete text must be fontified; and
* incomplete text must not break fontification of the following
  directive or transaction.